build-refactor-2: Fix build errors. Microkernel and kliblcd build.
Now that microkernel libcap build is separated from microkernel build
itself, when we trigger a top-level microkernel build, it doesn't
rebuild the whole thing every time. Kind of nice.

+++ lcd-domains/kliblcd/module_load.c
@@ -130,7 +130,7 @@ static int dup_module_pages(hva_t pages_base, unsigned long size,
 {
 	int ret;
 	void *dup_pages;
-	unsigned int unused;
+	unsigned long unused;
 	/*
 	 * Alloc dup pages, and memcpy the bits over
 	 */
